Rename module to better represent what it contains
This module provides an implementation of the RequestResponseCodec
using a cbor serialization.
