+++ swap/src/network/transfer_proof.rs
@@ -1,5 +1,5 @@
 use crate::monero;
-use crate::network::request_response::CborCodec;
+use crate::network::cbor_request_response::CborCodec;
 use libp2p::core::ProtocolName;
 use libp2p::request_response::{
     ProtocolSupport, RequestResponse, RequestResponseConfig, RequestResponseEvent,
